Add files via upload
A C++ program that computes all the maximal independent sets for a graphs of a certain size.
This program calculated the maximum number of maximal independent sets for graphs of size 1, 2, 3 ,4 ,5 ,6 ,and 7.
They are:
1 -- 1
2 -- 2
3 -- 3
4 -- 4
5 -- 6  (2*3)
6 -- 9  (3*3)
7 --12 (2*2*3)

The extremal graphs are also given.   They turn out to be the one's you expect,   For n=7, the extremal graph is K2 + K2 + K3.